bozo: Remove bozo_fileName global variable
Currently, the bozo_fileName global variable is always set to the path
of the BosConfig file and is never changed. A zero (NULL) value is
always passed to ReadBozoFile() and WriteBozoFile(), and the
bozo_fileName global is used within those functions to specify the path
of the BosConfig file.

Clean this up by removing the unnecessary bozo_fileName global variable.
Change ReadBozoFile() and WriteBozoFile() to expect a non-NULL argument
which is the file path of the configuration file to read/write, and
update the call sites to specify the dirname symbol for the BosConfig
file path.

Note that we keep the file path argument in ReadBozoFile() and
WriteBozoFile() to make it easier to write unit tests for those
functions in a future commit.

Change-Id: I8cdcadbcedbc7e7c0d4d3921724954205d55369c