cmd: Do not leak tokens in cmd_ParseLine()
The cmd_ParseLine() function first parses the input string into a list
of tokens, and then calls CopyBackArgs() which copies the character
pointers to the caller provided argv array.  An error is returned if the
provided argv array is smaller than the token list, however the excess
list elements are not freed.

Avoid this memory leak by freeing the excess linked list elements before
returning in this situation.

This memory leak was found by running the tests/cmd/command-t unit test
program with valgrind.

Change-Id: Ib4f8596ddfd87afbc0ed64b50cc7bf8bbfb8d2f7