RedHat: Add RPM_RELNOTES, RPM_CHANGELOG, and RPM_CELLSERVDB
Currently, when building the source RPM with make, the release notes and
change log are empty, and the CellServDB is downloaded from
grand.central.org.

Instead (by default) use the NEWS file for the release notes, generate a
change log from git (if a .git directory is found), and avoid
downloading the CellServDB file (since we already have a local copy that
matches the one in the spec file.)

Add the RPM_RELNOTES and RPM_CHANGELOG makefile variables so we can
override the defaults if we want to provide our own release notes and
change log files.  Specify empty values on the make command line to
fallback to empty files for the RELNOTES and ChangeLog.

Add a simple script to generate a ChangeLog from 'git log'. For now,
just use the output generated by git log, but avoid tag an branch
decorations.  Fallback to an empty ChangeLog if the project directory is
not a git repository.

Add the RPM_CELLSERVDB makefile variable to specify the local CellServDB
file to be packaged. By default, package src/afsd/CellServDB from the
working tree (which matches the one specified in the openafs.spec.in
used to build the source rpm).  Specify an empty RPM_CELLSERVDB on the
make command line to have makesrpm.pl download the CellServDB file from
grand.central.org.

Change-Id: Idf1afd1683d87d882d2c7e49058cbf18839e5c9e