cmd: Introduce cmd_Join()
Add the cmd_Join() function to combine an argument vector to a quoted
shell-like command string.

The number of arguments and the size of the arguments to be joined are
only limited by the available memory.  The returned command line string
is allocated by cmd_Join() and must be freed by the caller.

Change-Id: If569cb22e3a984d635c0b8c6e4f78c2980011391