man-pages: Add objdir support to merge-pod
The merge-pod script is our simple custom preprocessor for man-page pod
files. The merge-pod script expects one or more arguments to specify the
input files and generates the output files in the same directory as the
input files.  Unfortunately, this precludes us from using merge-pod to
do objdir builds which generate man-pages.

Change merge-pod so that when no input files are specified, merge-pod
will scan the man-page source directory for pod.in files, and put the
pod output files in pod<n> directories in the current working directory

With this change, merge-pod remains compatable with old method, which is
still in use by the NT makefile and the regen.sh script, but will allow
us to use it in the man-pages Makefile in the future to generate man
pages during the build.

Change-Id: I36b5b851cd1a09d050cf21c65ab3ae160a5c15cb