afsd: Unbuffer output
afsd prints output to the stdio streams. Usually afsd is run by an init
script or systemd unit, so the output is not directed to the terminal
and the stdio streams are buffered.

When afsd forks child processes, the forked processes inherit the
buffered streams and this causes the output to be duplicated in the
child proceses.  For example, when running afsd from systemd, the
journal contains duplicated output messages:

    # systemctl start openafs-client
    ...
    ... systemd[1]: Starting openafs-client.service - OpenAFS Client Service...
    ... afsd[292192]: afsd: All AFS daemons started.
    ... afsd[292170]: afsd: All AFS daemons started.
    ... fedora systemd[1]: Started openafs-client.service - OpenAFS Client Service.

To avoid the duplicated messages, and to ensure output is captured in
the event of a crash, set the stdout and stderr streams to be unbuffered
using setvbuf() when the afsd process starts.

Thanks to Mark Vitale for diagnosing this issue and to Andrew Deason for
suggesting the use of setvbuf() to set the stream buffering.

Change-Id: I57faac38f6667d431557793ce06a11b7f390a414